pyproject.toml
[project]
name = "connect4"
version = "0.1.0"
description = "Connect 4 game with Django backend and JavaScript frontend"
requires-python = ">=3.8"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "pydantic>=2.0",
    "django-cors-headers>=4.0",
    "tensorflow>=2.15.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

